project-medusa/medusa-cli/requirements.txt
# MEDUSA CLI - Python Dependencies
# Version: 1.0.0

# CLI Framework
typer[all]==0.9.0

# Terminal UI
rich==13.7.1
prompt_toolkit==3.0.52

# HTTP Client (async)
httpx==0.26.0

# Configuration Management
pyyaml==6.0.1

# LLM Integration
google-generativeai==0.3.2

# Report Generation
jinja2==3.1.3

# Web API Server
flask==3.0.0
flask-cors==4.0.0

# Database (for operation logging)
# sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
